--- a/Framework/Core/src/DataProcessingDevice.cxx
+++ b/Framework/Core/src/DataProcessingDevice.cxx
@@ -987,12 +987,19 @@ void DataProcessingDevice::handleData(DataProcessorContext& context, InputChanne
       switch (input.type) {
         case InputType::Data: {
           auto headerIndex = input.position;
-          auto payloadIndex = headerIndex + 1;
-          assert(payloadIndex < parts.Size());
-          auto dh = o2::header::get<DataHeader*>(parts.At(headerIndex)->GetData());
-          auto relayed = relayer.relay(parts.At(headerIndex),
-                                       &parts.At(payloadIndex), dh->splitPayloadParts > 0 ? dh->splitPayloadParts * 2 - 1 : 0);
-          ii += dh->splitPayloadParts > 0 ? dh->splitPayloadParts - 1 : 0;
+          auto nMessages = 0;
+          auto nPayloadsPerHeader = 0;
+          {
+            // multiple header-payload pairs
+            auto dh = o2::header::get<DataHeader*>(parts.At(headerIndex)->GetData());
+            nMessages = dh->splitPayloadParts > 0 ? dh->splitPayloadParts * 2 : 2;
+            nPayloadsPerHeader = 1;
+            ii += (nMessages / 2) - 1;
+          }
+          auto relayed = relayer.relay(parts.At(headerIndex)->GetData(),
+                                       &parts.At(headerIndex),
+                                       nMessages,
+                                       nPayloadsPerHeader);
           switch (relayed) {
             case DataRelayer::Backpressured:
               if (info.normalOpsNotified == true && info.backpressureNotified == false) {
@@ -1142,8 +1149,8 @@ bool DataProcessingDevice::tryDispatchComputation(DataProcessorContext& context,
     auto getter = [&currentSetOfInputs](size_t i, size_t partindex) -> DataRef {
       if (currentSetOfInputs[i].size() > partindex) {
         return DataRef{nullptr,
-                       static_cast<char const*>(currentSetOfInputs[i].at(partindex).header->GetData()),
-                       static_cast<char const*>(currentSetOfInputs[i].at(partindex).payload->GetData())};
+                       static_cast<char const*>(currentSetOfInputs[i].header(partindex)->GetData()),
+                       static_cast<char const*>(currentSetOfInputs[i].payload(partindex)->GetData())};
       }
       return DataRef{nullptr, nullptr, nullptr};
     };
@@ -1261,9 +1268,9 @@ bool DataProcessingDevice::tryDispatchComputation(DataProcessorContext& context,
       int cachedForwardingChoice = -1;
 
       for (size_t pi = 0; pi < currentSetOfInputs[ii].size(); ++pi) {
-        auto& part = currentSetOfInputs[ii][pi];
-        auto& header = part.header;
-        auto& payload = part.payload;
+        auto& messageSet = currentSetOfInputs[ii];
+        auto const& header = messageSet.header(pi);
+
         if (header.get() == nullptr) {
           // FIXME: this should not happen, however it's actually harmless and
           //        we can simply discard it for the moment.
@@ -1271,12 +1278,12 @@ bool DataProcessingDevice::tryDispatchComputation(DataProcessorContext& context,
           continue;
         }
 
-        auto fdph = o2::header::get<DataProcessingHeader*>(header.get()->GetData());
+        auto fdph = o2::header::get<DataProcessingHeader*>(header->GetData());
         if (fdph == nullptr) {
           LOG(ERROR) << "Data is missing DataProcessingHeader";
           continue;
         }
-        auto fdh = o2::header::get<DataHeader*>(header.get()->GetData());
+        auto fdh = o2::header::get<DataHeader*>(header->GetData());
         if (fdh == nullptr) {
           LOG(ERROR) << "Data is missing DataHeader";
           continue;
@@ -1303,15 +1310,19 @@ bool DataProcessingDevice::tryDispatchComputation(DataProcessorContext& context,
 
         if (copy) {
           auto&& newHeader = header->GetTransport()->CreateMessage();
-          auto&& newPayload = header->GetTransport()->CreateMessage();
           newHeader->Copy(*header);
-          newPayload->Copy(*payload);
-
           forwardedParts[cachedForwardingChoice].AddPart(std::move(newHeader));
-          forwardedParts[cachedForwardingChoice].AddPart(std::move(newPayload));
+
+          for (size_t payloadIndex = 0; payloadIndex < messageSet.getNumberOfPayloads(pi); ++payloadIndex) {
+            auto&& newPayload = header->GetTransport()->CreateMessage();
+            newPayload->Copy(*messageSet.payload(pi, payloadIndex));
+            forwardedParts[cachedForwardingChoice].AddPart(std::move(newPayload));
+          }
         } else {
-          forwardedParts[cachedForwardingChoice].AddPart(std::move(header));
-          forwardedParts[cachedForwardingChoice].AddPart(std::move(payload));
+          forwardedParts[cachedForwardingChoice].AddPart(std::move(messageSet.header(pi)));
+          for (size_t payloadIndex = 0; payloadIndex < messageSet.getNumberOfPayloads(pi); ++payloadIndex) {
+            forwardedParts[cachedForwardingChoice].AddPart(std::move(messageSet.payload(pi, payloadIndex)));
+          }
         }
       }
     }
@@ -1319,7 +1330,6 @@ bool DataProcessingDevice::tryDispatchComputation(DataProcessorContext& context,
       if (forwardedParts[fi].Size() == 0) {
         continue;
       }
-      assert(forwardedParts[fi].Size() % 2 == 0);
       // in DPL we are using subchannel 0 only
       device->Send(forwardedParts[fi], spec->forwards[fi].channel, 0);
     }