DataRelayer is using new version of MessageSet API which will
allow to switch message store in MessageSet transparently

Continuation of #7110 

There is a speedup in the message handling in DataProcessingDevice for split parts

Before
```
Run on (8 X 3400 MHz CPU s)
---------------------------------------------------------------
Benchmark                        Time           CPU Iterations
---------------------------------------------------------------
BM_RelayMessageCreation        193 ns        193 ns    3580395
BM_RelaySingleSlot             987 ns        986 ns     705237
BM_RelayMultipleSlots         1012 ns       1012 ns     661466
BM_RelayMultipleRoutes        1880 ns       1880 ns     374791
BM_RelaySplitParts            3806 ns       3792 ns     190106
```

After
```
Run on (8 X 3400 MHz CPU s)
---------------------------------------------------------------
Benchmark                        Time           CPU Iterations
---------------------------------------------------------------
BM_RelayMessageCreation        199 ns        199 ns    3567992
BM_RelaySingleSlot            1000 ns        999 ns     700044
BM_RelayMultipleSlots         1012 ns       1012 ns     683663
BM_RelayMultipleRoutes        2012 ns       2012 ns     345065
BM_RelaySplitParts            1391 ns       1390 ns     521503

```